chore: move timer controller to another file
