+++ lib/features/start_workout/start_workout.dart
@@ -1,3 +1,4 @@
 export 'cubit/cubit.dart';
+export 'utils/timer_controller.dart';
 export 'view/start_workout_page.dart';
 export 'widgets/widgets.dart';
