+++ lib/features/start_workout/utils/utils.dart
@@ -1,0 +1,1 @@
+export 'timer_controller.dart';
